LUA - AI Outnumbered modifiers
There was some confusion in how the 'outnumbered' condition impacted the AI - and the magnitude of the impact.   

Moved from using the full cheat value, down to using the MajorCheatModifier (65%), to modify the base cheat value for an outnumbered AI.   This also cleans up the mechanics of the maintenance modifier, making it the same value.