LUA - AI Starting Mass Point list
One way that we keep LOUD moving forward, especially early, is by giving him a list of mass points to focus upon.   This does wonders when you have a team of AI - since they don't fight over the initial mass points.

Previously, we limited this list to about 11 entries - which is sufficient on smaller maps, or games with a large number of AI - but we wanted to adapt it for use with an outnumbered scenario - so that an outnumbered AI will have a larger list.

We increased the base minimum for all, and added a multiplier for an outnumbered AI.   We also allowed the limit to float per AI - whereas it was previously just fixed.