LUA - AI Transport building
Streamlined the Transport building process to be entirely driven by need.   That need is generated by both engineers and platoons that call for transport.

All 3 levels of transport will now only be built if and when demand exists.

Some further refinement to TransportUtilities module, bringing the creation of the TransportPool into the module and removing it from the ScenarioUtilities.