LUA - PARSEINTEL flaw
Found a serious procedural flaw in the way I was building the HiPri list for the AI.   The result was that in some conditions, the threats were being build properly - but the location of the threats was not.  

This was just a rookie error on my part about using reference data - and not real data.

Aside from that, we did some study on threat mechanics and I have introduced a new threat type that we will monitor 'AntiSub' - to better follow underwater threats.