LUA - AI BaseMonitor
Rebuilt the structure of the BaseMonitor cycle.   

Previously I used a copy of the HiPri table, but this table has data not used in BaseMonitor, and includes threats far beyond the range of the base.

We'll build a much more compact list out of HiPri by pre-filtering it for threat types and distance - and then we'll use that much shorter list - and only process the types of threats that we found when we built the list.

Net result - the BaseMonitor frequently has ZERO things to process, resulting in substantial increase in throughput, and less overhead.