UNITS - Reverted Cybran Rhino laser effect
Since there was naught but complaint, I have altered the Cybran T2 Rhino to use the same beam as the Cerebus T2 PD.    In fact, the weapon is now the same, doing 10 damage per tick, per barrel, albeit with a much shorter firing cycle of 2 ticks, twice per second.

This increases the DPS from 72 >> 80 - but there is a commensurate E cost on the unit to go with it.