LUA - AI Engineer fix
There was an edge condition that was causing engineers to 'stall'.   This was being caused by the difference in 3D positions of engineers, versus the 3D format (but 2D data) of build positions.

This would cause the engineer to think he was still very distant from his build location, when in fact he was right on top of it.   

This condition was exposed on those maps where the vertical terrain layer is much higher than the base value ( > 100 ogrids).    Those maps are rare, and that kind of elevation difference is extreme, so it shows up quickly, but it's likely something that's been around for a long time.