T4 Naval Units - BO, BrewLAN, UNITS
T4 Naval units get an energy review (ala artillery) and the range of most weapons is now in line with tactical artillery, which has resulted in a minor decrease in range for many of the big guns.  Missiles are unaffected.

I have reviewed the firing sequencing of most of the big guns as well, and confirmed their DPS values so that the blueprint and the spreadsheet data align.

Most of the big ships required their acceleration and turn parameters to be reviewed as well, with most now taking 4-5 seconds to reach full speed.

Unit cap costs have been moved up, to be more inline with the number of weapon systems on the unit in question.    They still represent a good value in that area - but they were getting off very light previously.

The drones on the Tempest were not being costed fairly (super cheap and almost immediately replaced on loss) so they've been costed in-line with gunships, which is pretty much what they are.

The Seraphim T4 Cruiser has been more appropriately positioned, as a LIGHT naval experimental, with reduced HP and associated costs, this unit should be viewed as a T3.5+ Cruiser rather than any kind of surface monster - it's primary advantage is superior AA weaponry. 

AA weaponry, in the T4 Naval class, has been brought more inline with it's T3 equivalents.

The UEF Dreadnaught has had it's firing animations brought in-line with it's firing cycle - so the state of the reload and readiness to fire can be visually monitored with the barrel recoil.