LUA - Lobby options, NoRush issue
I have reversed the addition of the Sera Nuke Battleship to the NUKE restrictions, and likewise, all the NUKE subs are also removed.    The code should detected the NUKE weapon and disable that specifically.

This does not, of course, address the hideous impact this will have on the investment into these units, since that weapons cost is part of the unit's overall value.

A recent issue with the NoRush timer, on a very small overcrowded map, recently revealed itself.   This minor patch keeps the AI from doing any platoon formation (which was the cause of the issue) until immediately before the timer expires.   We'll continue to monitor this issue.