LUA - Weapon Restrictions
It's interesting to note that specific weapon restrictions are possible, and that the code was always in place - BUT - the data to make it happen was not there.

Specifically, NUKES and TMLS have a method of restricting those specific weapons while keeping the base unit in the game - but - the data necessary to make that happen didn't exist.

It does now - and by doing so - we can restrict nukes and tac missiles by marking the necessary weapons and selecting the restriction