LUA - DEBUG - weapon state logging
Added more detail to weapon state logging so it's easier to follow weapon rigging and the various delays imposed over the course of a firing cycle.

Remove the limitation that prohibits creating weapons with a RoF that would exceed the muzzle delay period.   Doing so doesn't create an issue - the weapon will just 'skip' that OnFire event.