LUA - Teleporting and UI display
During teleportation, the economy values of the unit being teleported are being incorrectly shown - being based off the normal consumption and not any modified values.   The internal values are correct.

The bug is most visible after the teleport is complete, where the real values are not being properly restored.  By forcing the Consumption to Active, the proper values will be reinstated on the UI.