PROJECTILES - Ver 6.3.01 LOUD
This a large data reduction pass on the LOUD projectile blueprints.   It accomplishes a number of things.

In most cases, the factional category and other, redundant categories have been removed.

The biggest memory gain in this process comes from the almost total removal of the 'General' data block for almost every LOUD-based projectile blueprint.  In testing, this resulted in a general reduction of memory usage of a full percent.   Not much in the grand scheme of things, but I'll take 5MB of additional workspace anyday.

I have also taken the time to install maximum shooter limits on things like torpedoes and tactical missiles.   This should substantially reduce overkill on those targets - since it will keep multiple defense systems from all firing on the same target.