LUA - AI ACU Defense Builders
We want the ACU to participate in defense building for T2 and T3.   These builders allow him to do this work, on his own, at the appropriate stage.

We had to create a function to acknowledge his ability to build T2.

We made an adjustment to a trigger point used by some other builders.

Create 2 new builders, one for T2, and another for T3 - when both are active - the T3 build will take precedence.