LUA - PHXLIB - threat override calcs
A flaw was discovered in our threat logic that allowed some units (mainly air), to exit the threat calculation prematurely - and have ZERO threat assigned to them - if the code encountered a weapon with No Rate of Fire (Air Crash weapons for example).   Properly, the code should have just continued on, but was instead returning NO RESULT at all.   Fixed.

We also installed a gate for those weapons that are isolated to AIR targets only, from passing a minor portion of their threat to surface attack.

I have cleaned up the debug entries and organized the reporting so that it more clearly identifies the process of the code.  

I've installed some logging to identify just how many units are being threat mapped, and the difference between the blueprint and the new value.