LUA - New AI metric - Production focus
We're developing a new metric for LOUD to better understand the effort of his opponent - Production focus.

While 'factory counting' has previously been used in other AI, it's never had any logic other than simple counting applied to it.   In this metric - we assign a value to the tech level of each enemy factory - and then use it to arrive at a value that gives LOUD some idea of the emphasis that his opponent is taking.  

Aside from just knowing about the presence of the factories, LOUD will attempt to gauge if the factories are being utilized or not - with non-utilized factories being discounted.

This may be a valuable metric in helping LOUD to understand the production focus of his enemy - and adjust his own stance as a result. 

This is just the construction stage - we have not yet implemented the result in any active manner.