LUA - BUFFS
Diagnosing a problem with the mostly un-used weapon buffs revealed an oversight with beam weapons not being affected by them.

In review, the buff system has no mechanic to identify a specific weapon, and only affect that - instead it will buff ALL the weapons on the unit.

I have created a method to specify which weapon # to affect, in the buff, and make that a necessity - but I have commented it out for now.