LUA - WEAPONS logging, beam creation
This review is primarily about beam weapons, the creation process and sequence, and some logging that was missing, for beam based functions.

The first part of this process was following how beam weapons are created, and what sets them apart from other projectile weapons.

Aside from the weapon itself, beam weapons have an extra creation process which creates the beam entity which the weapon will use when it fires.  I have re-arranged the sequence of this creation, to put the beam creation AFTER the weapon creation, and log it accordingly.

The actual beam, that we see, is an effect, that occurs when the beam connects from one point to another.  To do this, the beam must be Enabled() and then it's start and end points are declared.

The process begins via the CreateProjectileAtMuzzle() event, which is common to all weapons when they fire.  For beam weapons, this process iterates over all the previously defined beams - and, if not already enabled, the PlayFxBeamStart() function is called.

The PlayFxBeamStart() process will Enable() the beam, and startup any lifetime and audio threads.
The OnEnable() event, in CollisionBeam, calls CreateBeamEffects().

CreateBeamEffects() creates the visuals.