LUA - CollisionBeam Impact event
The OnImpact() event for beams has some differences versus the typical Projectile OnImpact() event.   The differences stem from the structure of beam weapons, and the creation process of the beam entity. 

The beam entity was originally getting it's damage data directly from the blueprint, every time this event occurred - Projectile weapons, OTOH, create this data once, and simply re-use it - and buffs can modify that data and it will be persistent.

For beam weapons, they too created this weapon based damageTable, and buffs were indeed changing it, but the OnImpact() event just didn't use it, rebuilding it instead from the blueprint.

We now retrieve damageTable from the weapon (and not the blueprint) and weapon buffs now work like Projectile weapon buffs.