LUA - Relocate Starting Resources
We apparently had a logic issue with this, and it was worth having the discussion about what we really desired to see.

In the end - we arrive at two outcomes.

On - ALL mass & hydro at ALL start locations are relocated - including HUMAN starts.

Off - Mass & hydro at HUMAN start locations are NOT relocated.