BrewLAN for LOUD - Ver 0.8.8.9.70
A general review of the Penetration fighter series, this is a general performance pass, which moves the traverse range of the weapons on all of them, upwards, to be somewhat more competitive with the T3 ASF.   

All Pen Fighters are now 2 Unit Cap each

Increased Regen on all  (+4 to +6 improvement)

Dynamic Collision sphere, when not travelling at full speed, the collision sphere will decrease to 60%, which is much more in-line with what is needed at those lower speeds.

Just in general, the pens will perform much better in an A2A role than previously.