LUA - Improved Attack Plan dialog
As I continue to develop the data for the new PRODUCTION FOCUS metric, I'm taking the time to improve some of the logging functions associated with it.

The visual Base Monitor has been moved so that it always shows, if turned on.   This changes the behavior which previously would only appear if there was a threat nearby.