LUA - Collision Beams
There are a handful of beam weapons that will drop a 'splat' decal when they do an OnImpact() event.   The thread, which monitors this, is housed in the individual collision beams.   This thread was attempting to stuff itself into a Trash object that was previously housed in the 'beam' entity itself.

We have eliminated that Trash, and all trash is now housed on the weapon entity, that houses the beam.

Reduced the size, duration, and some of the LOD levels for these splats.