LUA - Disperse to Rally Points
This function redistributes units to rally points around a base.   In it's base form, it will take a list of units and simply direct them randomly to all the points generated by a given base.

In it's advanced form, you can specify a given point, and how many points you wish to make available - and the available points will be sorted by proximity to, and the number of points - allowing this function to concentrate units in or around a given side of a base.

We use this function when platoons RTB, but it's also important in the BASE MONITOR process, positioning units closer to inbound enemy threats.

The advanced functionality has now been distributed to a number of places, including, and most importantly, to the RTB of distress response formations when a base alert has been executed.