LUA - BASE MONITOR, DISTRESS RESPONSE
The BASE MONITOR process detects threats inbound and around a given base, and directs units to respond to those threats.

The detection process evaluates threats from IMAP for proximity, and then analyzes them for detailed composition using a radius of 120 ogrids - if necessary a BASE ALERT is created.

The raising of a BASE ALERT can trigger a DISTRESS RESPONSE, and the distress response was not using the same method to track the threat.   We have improved the threat tracking, to utilize the 'rings' equivalent value of the 120 ogrid range.

I have also formalized the log message formats.