LUA - Advanced Tracking for anti-projectile weapons
As we move to a better anti-torpedo system, it is necessary to activate the 'retargeting' ability that we currently have for unit seeking missiles, and make it available to projectile-seeking munitions.

This is not as simple as it sounds, since retargeting a unit-seeking munition is, by comparison, much easier to do - just hunt for another unit.

Hunting for projectiles is not so easy - you cannot just 'look' for another projectile in the area.  Instead, we need to rely upon the weapon which fired the munition, and go after whatever it's current target may be.   In most cases, the weapon itself will have hit the maxShooter cap on the original target, and moved onto another target (and may even have already fired upon it).    However, that circumstance is just fine.

To do this, when we detect the creation of an 'advancedTargeting' weapon - we test if it's one that seeks projectiles, and if it is, we store a reference to the weapon, on the projectile damage table.   In this manner, when retargeting is necessary, the projectile can reference back to the firing weapon, and get a new target, if any.