UNITS - Ver 6.3.10
This UNITS update dovetails not only with the removal of the final CIVILIAN units, but a great many data reductions across almost every unit.

Also included here, is the addition of 'advancedTracking' to a number of the anti-torpedo systems that use discrete projectiles.

This pass includes the final work on all AIR and NAVAL units, and some minor (but unfinished) work on all others.