LUA - Silo weapons
Silo weapons, both static and mobile, usually function on a 1-1 basis.   The BrewLAN mobile UEF anti-nuke is unique in that it requires multiple missiles for a complete salvo.

This was causing a malfunction where not only would the unit not fire, if it didn't have enough missiles, it would lock the inbound missile, and prevent any other unit from shooting at it (based on the projectiles MaxShooter value).

It was clear, from testing, that this lock would occur as soon as the weapon saw a target, if it had ANY number of missiles ready to fire.    The only way to prevent this lock was to drop the weapon range to 1 - so that it couldn't acquire, or hold, a lock.  

It was necessary to understand the available callbacks, and utilize the, to get the unit to watch it's ammo count, and it's muzzle salvo size - and dynamically adjust the weapon range as needed.

I had originally attempted to utilize weapon states to get a hold of this process (completely unnecessary in the end) - and to that end, I have expanded a lot of the weapon state logging events.

I have re-enabled some of the events associated with this, and created 2 new events, to signal ammo reductions.   Those new events (and the existing ones) are triggered within a monitoring thread that will watch the ammo on the unit.