LUA - Defaultweapons - reload animations
Reload animations, and the RackSalvoReloadTime are very closely related - unfortunately, the two are mutually exclusive - some units have this reload time, but no animation, while some have a reload animation but no delay value.

Ideally, the delay - and the animation - should run concurrently - but they do not.   The timing of the animation is also an issue, since it's not immediately clear how long that might be - and, of course, the animation rate can be adjusted - which makes it even more unclear.  That's another subject though.

For now, we'll drop into RackSalvoReloadState if EITHER is present.