LUA - Default anti-projectiles, Projectile
The FLARE entity was not being passed the target category in a proper fashion 

The FLARE now features a 'growth' process in the size of the collision sphere, and is properly added to the trash of the owning projectile.

The projectile itself, used to track the target missile, and would destroy when that missile died.   This is no longer the case, the 'bubble' entity will survive it's full lifetime unless it is impacted by a missile.  This applies specifically to the AEON TMD anti-missile