BO Unleased for LOUD - Aeon ARIA T3 PD
The weapon was experiencing some targeting issues.

Set LeadTarget to false, and decreased the FiringTolerance, which forces it to aim more accurately.   Removed unnecessary projectile values (it's a beam after all).

Reset the rotator, on the back, to speed up when firing, and stop when off, while emitting some weapon steam when stopping.

A more interesting addition - the weapon will now elevate (to it's normal height) when firing - and retract to a lower profile when stopped.   This is just me being creative with some effects and bone manipulation.  The collision box has also been slightly reduced in height.

I rarely have the chance to play with those things, so this seemed like an easy entry point.  I hope you like it.

Added a 2nd targeting bone, which will aid in making the unit harder to hit, when retracted, and behind walls - not impossible, but harder.