LUA - DefaultWeapons, animation timings, delays
I have harmonized the animation and delay handling associated with weapon scripts.   In all cases where both an animation, and a delay are in use - the animation play time will be deducted from the delay time.   This allows us to properly predict the firing cycle, in conjunction with animations - so that the threat logic can properly valuate units.

I have greatly expanded the Weapon and Weapon State AI Debug logging so that it is now possible to see the length of the animations on the units, and judge the timing impact.