LUA - DefaultWeapons
There are a lot of unusual decisions in this process, and the many states that weapons go thru are often a mystery to most - I have made every attempt to LOG all these weapon state changes, and any timings or delays associated with them.    Use the LOUD AI Debug tools if you ever want to look into the full details about a weapon and it's firing cycle.

Perhaps the most interesting discovery comes around weapons that use Energy to fire.  The CreateEconomyEvent() process takes 3 values - Mass, Energy and Time.    This has long been used and it's reliable and stable - BUT - guess what ?  It always adds 2 ticks to the actual time required.   In most cases, it's not used for timing critical things, so we don't notice this, but in Energy Weapons, it's very time critical.  I have taken the step of deducting these two ticks from the time required for this process, with weapons.   This permits the user to see accurate timing values in his blueprint.

The sequencing of the firing states is difficult enough to follow as is, and there are many shortcut situations - this work should help future users identify both timing and operational issues, and it will help identify the many relationships between not only the visuals (animations) that many weapons have, but the outcome.