LUA - DefaultWeapons
Some additional commenting and refinement.

Apparently, the math functions like to turn what should be integers into floating point values, resulting in some timing issues, again, with the Economy Events.   To avoid this, it was necessary to add a 1.001 multiplier, to the event time, to insure tick level accuracy across the normal gamut of values.

I have converted all timings over to actual ticks rather that time, which was subject to the issue above.

Further refinement of the logging to better denote delays that do not impact the firing cycle versus those times that they do.