TM - Data reduction, timings, animations
A wide number of units got data reductions for the use of 'RackSlavedToTurret' which is only required if  you have multiple rack bones that needed to be tied to a single pitch bone or a pitch bone which is different from the rackbone.   It is also not needed if you have dual manipulators specified.

More of this will be coming down the line as I work thru the TM units.