LUA - Attack Plan management
The Attack Plan is responsible for the focus of LOUD's attacks, so it's important that it accurately recognize and evaluate threat positions - and shift the focus accordingly.

The Attack Plan position is responsible for setting the production focus of a base, either Land or Amphibious - in order to maximize unit efficiency for the current situation.   Due to some bad work, in picking what the most threatening position, some bases were switching to an 'Amphibious' state, even on all land maps - because they could not find a legitimate land path to the selected position.

This also fixes a condition where the builders were not being properly resorted when a base changed it's PRIMARY status, resulting in the creation of reinforcement platoons that were being told to go to the base they were already at !

Some efficiency in the selection process - we no longer process Naval threats on maps that don't have naval markers