LUA - Builder Priority Function handling
When a builder changes priority from it's default (starting) value, the original value is stored in the OldPriority variable - and when reverted to it's original value, this variable is cleared - allowing it to act as not only a semaphore to indicate the builder has been altered - but what the the default value is.

I wasn't considering that use in a lot of Priority functions, and as such, once changed, some builders were never able to reference the basic normal value.  This allowed some builders to persist, thru a change into or out of PRIMARY status - when they should have been disabled by such a change.

Most of these changes revolve around retrieving the OldPriority first (if present) before using the current priority (which was assuming the original value).