LUA - Transport Utilities
GetTransports() will now return a 3rd result - the airthreat of the transport platoon.   This is used to provide more accurate path finding results, based upon the real values coming from the transports.

Also, we properly return the number of transports being used - previously it was returning the unit count of the unit platoon.

The main side effect of this is the proper use of the land units existing path - when a simple transport is too dangerous.   This allows the transport platoon to deposit a platoon at a safe point along it's existing travel path - if the final destination cannot be safely reached.

Cleaned up some of the debug dialog