LUA - Attack Planner
The Attack Planner generates a plan of progression to reach the most suitable enemy by identifying the enemy position with the most threat - and building a series of steps towards that.

Heavily impacted by distance, the previous calculations were not accumulating multiple threat types from the same IMAP location - instead, only responding to the one with the most threat - with the distance modifier applied.

This sometimes led to missing the most obvious, and usually closer, threat positions.

A second issue, has to do with point selection, with the planner not accepting a point unless it was 30% closer to the goal, than the current point.   This has been relaxed to 25% closer, resulting in more reliable point selection.