LUA - AI PlatoonToPositionDistanceTriggerThread
This thread controls the standard movement of platoons from point to point.

When movement is taking too long to reach the waypoint, the platoon is ordered to move to the waypoint in a 'loose' fashion.   This is working, but - a delay of 10 ticks was being added for EACH unit - instead of just once.

The delay period between checks has been relaxed a bit, reducing work load.

Improved logging.