LUA - AI Attack Planner
Substantial rewrites to the attack planner for improved memory and more complete data structure

Improved logging of selection process

Added use of 'AttackPlanner' in the platoon name when doing path finding operations.   This takes advantage of a new feature in the pathfinder which allows bypassing terrain blocking.

Resolved some issues with pathfinding and improved overall logging.