LUA - AI GetHiPriList & RecheckHiPri
The AI utilizes a filtered set of data from the HiPri list when looking for suitable targets.

Added ability to feed the required threat types to the function directly.   Also I now feed it the max distance, and a control of if we perform threat valuations on Everything we find in the area - or only those indicated by the threat type.

GetEnemyUnitsInRect() function now reports the number of targets in the results