LUA - AI Pathfinders
Adding some new features, the ability to bypass threat checking by use of 99999 threat maximum (used by Attack Planning) or other non-platoon based pathing checks.

DestinationBetweenPoints no longer looks too far beyond the destination and no long starts as close - this reduces the workload.

Threat calcs are now entirely built on GetThreatBetweenPosition BUT - the value is modified by the distance of that step - with nominal being the IMAPRadius ( half the width of an IMAP block ).   This adjusts the threat down when the trip is  short - and up if it's a long one - based upon the length of the distance from point to point.

Amphibious movement now more greatly penalizes land movement over water.

Blocking check can now be bypassed