LUA - AI RetreatAI
RetreatAI monitors the status of a platoon and retreats it if the platoon strength falls too low.

This function relies upon being notified of the platoon being 'Under Attack' - which, originally, was only responding if the platoon was running a DistressResponseAI() behavior.

The 'UnderAttack' status will now be triggered if the platoon has either function running.

RetreatAI will mark the platoon as such.

RetreatAI will now measure which type of strength the platoon is mostly made up of - and retreat when that value falls too low.   Previously, we were only responding to the OVERALL strength, which was inadequate when the platoon had lots of support units (AA for example) left - but the main body was already dead.

Reduced the timing on the response thread when the platoon is UnderAttack, from 3 ticks to 2.