LUA - Ship/Submarine Sinking effects
One of the biggest contributors to the 'Particle Cap exceeded' log messages appear to be the sheer number of water effects going on when a ship sinks.

I have simplified a portion of this, in an effort to reduce that minor issue - but also to make ship wrecks reliably appear on the seabed.

Previously - and this is an animation issue - the sinking process was being governed by the animation - often lasting too long - and sinking below the seabed completely.    You would never see the wreck - but it was there.   The other nasty side effect, some animations are WAY too long - I have increased the default animation speed - if not specified - to 2.

Also - there was a rather silly surface explosion of debris, when the final wreck was created, on the seabed - I have removed that effect by passing a flag along to the 'DeathThread', to suppress that event - and the use of TryCopyPose() - all seabed wrecks will now sit flat on the seabed.