LUA - AI AmphibForce/LandForce
These two central platoon functions had a wee problem - platoons that were air transported, could go into a 'stall' if, when they landed, they couldn't path to the original goal.

Now, when pathing cannot be completed, the platoon will abandon it's current targetlocation, and being a new targeting cycle.

Some important changes to how targets are valued has been done.   A much closer examination of economy value, and threat values has given us a better method of evaluating targets.   The military value - and the economic value are now much less likely to 'swamp' the calculation - resulting in a much narrow range of valuations - and one that is more even weighted between the two.    This is especially true in factoring in the threat - along each step towards the target - based upon not only the basic threat - but the length of the step, thru that threat.

One other change is how often we call for transports - I have reduced it slightly - and the frequency of the HiPriority validity check.