TM - Sera T2 Tactical Bomber
This one-off unit, has some rigging issues - and was making use of a rare weapon switch which allowed the weapon to bypass several important weapon states.

With the cleanup of the weapon states in general, this 'shortcut' became obvious - allowing the weapon to begin firing bombs in an endless stream - as it was bypassing the reload state.

This was lazy work on the part of the author - all that was really needed was a reload delay.   This is not all that uncommon, as many authors are not conversant with all the weapon options.    In this case, the author likely didn't know how to get his weapon to fire in the time he liked, so he just used a 'hack' instead.   Sorry - that hack don't work no more.