LUA - AI AirBias
This metric, built from the air threat data, helps the AI to understand the nature of the enemy air strength by splitting air-to-air threat, from air-to-ground threat - and then comparing the two values.

This gives the AI a decent idea of the composition of the enemy air force, either fighter based, or bomber/gunship.

We'll use this bias to adjust production and behavioral reactions in some future work.