LUA - DefaultWeapons
Energy weapons have always had a blueprint function that bypasses the charging for the first shot.   This same variable can be re-used to indicate when a weapon has successfully been charged - and is ready to fire.

This permits getting rid of some minor threads that were intended to monitor firing beams, and shut them off in an out-of-power condition.