LUA - Collision Beams
A variety of work here - first up, the creation of 'scorch' splats.   Previously, this was being handled by a snippet of code in individual beams - now - we detect the presence of the data, and trigger some common code.   This allows me to remove all of that custom code, and drive it with just 3 pieces of data - SplatTexture, ScorchSize and ScorchTime

This impacts ALL unit mods that have this function - removing a great deal of repetition, and it allows me to remove a large set of locals used in DefaultCollisionBeams