LUA - Weapons - Trash handling
We've been using the Weapon trashbag to store all manner of things for a long time - but it's clear it's not the most efficient way to do this.

A number of things are repeatedly added to that trashbag, pushing the counter to very high values - in this update - we isolated, moved and eliminated several trashbag items for better memory allocation, and to eliminate some redundancy.

A great deal of improvement to the Weapon and Projectile Debug dialogs is also included, with an emphasis on expanding the beam dialogs.

A fuller understanding of how Beam weapons function, in general, and how that differs from the typical Weapon >> Projectile relationship was quite useful in cleaning up the data used in the creation and usage of beam weapons, in particular, the use of the Beams table to carry re-used data, such as the damageTable, and the BeamMuzzle.

Several of the Weapon States have hooks to events like OnFire(), etc - that don't need to be logged - now that I have a clear picture of their overall purpose, and usage.