LUA - Advanced Tracking
We've had this feature a long time, but we recently expanded it to encompass some high-end torpedo weapons.

We've discovered that it's original AA lineage was causing torpedoes to re-target on air units rather than targets that were valid for the weapon they were launched from.   By adding the handle to weapon, to these projectiles, we can re-target based upon what the weapon is tracking - which is much more sensible - and corrects this misbehavior.