LUA - DeathWeapons and buffs
The process for DeathWeapons has always been somewhat convoluted.

In many cases, DeathWeapons are NOT regular weapons at all - and don't create an entity - which is efficient - and has reliably triggered deathweapon damage, in a radius.    However, DeathWeapons were NOT processing buffs.

Some units have a 'death buff', like the Cybran Loyalist, and in order to make this happen, the DeathWeapon had to be custom coded - not only as an active weapon (which created a weapon entity), but some tricky script handling, to make it fire, and do it's buff before the weapon itself was destroyed by the unit's death event.

This was messy - and inefficient, and recently we've uncovered it was less than reliable, since if the weapon is destroyed quickly enough, the buff cannot even fire.

No - the proper solution was to find a way to not only move the buff, off the unit, and onto the weapon itself.   But to make the weapon a typical DeathWeapon.

These changes do just that - giving buff ability to the DeathWeapon - without needing to create a weapon entity, or actually go thru a firing sequence.    Now we can not only do area damage, but all manner of buff events - all of them contained in the weapon blueprint, just as with other weapons.