UNITS - Cybran T2 Amphib tank
This unit had some script that was disabling weapons based on a LayerChange() event - problem is simple - once disabled via this, it wasn't being re-enabled properly.

The base weapon code handles this anyhow making it redundant.

Some weapon review in order as well, the default shield damage was too high, and the rocket velocity and aim bones were wrong.