LUA - torpedo effects
Many torpedoes were playing a surface plume effect, even if detonating well underwater - this was just laziness in the data, since the code knows the difference between a hit underwater and a hit against a surface target.

The removal of the surface plume should ease the particle load on large torpedo interactions.