LOUD_MISC - AI DEBUG
It's taken some time for me to get to this, but the AI Debug panel will now show the status of the debug switches when loading into a new game or during a replay.

I've also addressed the DispalyIntelPoints function by showing the colors assigned, and the proper ability to toggle their display during a simulation (not a replay).

Organizing this data has allowed me to remove redundant data in ParseIntelThread - and insure that the colors are assigned, and maintained, entirely by the AI Debug module - which will feed them, via a callback, to the SIM, for display - instead of needing to maintain them in both locations.