LUA - Weapons - Energy Weapons
The WeaponCharged semaphore was being set by the EconDrain thread, when complete - thing is - that thread would signal termination based upon the end of the Economy Event - and not the completion of the script.

In some cases, the weapon would be charged, but not be signaled as such, and another charge cycle would start - delaying firing again.

I have moved the semaphore and duplicated it to all those places that are waiting on the EconomyEvent to end to insure the signaling is completed.