LUA - Weapons
An interesting script bug from a weapon on a unit was passing UNIT information, and not WEAPON information, causing the script to crash when fired (Anachronon).

Since the function, that was crashing, is reserved for use only with weapons that have the WeaponIsEnabled flag set, this can act as a guard from this situation.

The unit itself, will be fixed later, but this change is not only good for that, but slightly more efficient as well.