BrewLAN for LOUD - Hades T3 PD
The Hades is a Cybran T3 Energy based PD with some very unique features.

Unfortunately, out of the box, some of those features are at odds with changes we've made to the weapon sequences - and I wanted to rectify this.

The weapon has a basic energy consumption that increases with each shot taken, and a cooldown that reduces this demand over time.

If the weapon fires continuously, long enough, it will overheat and cause damage to itself.   We now visually indicate this with text appearing over the weapon.   Eventually, if the weapon is not shutdown, this damage will destroy the weapon.